scripts/requirements_mysql.txt
requests==2.31.0
faker==20.1.0
mysql-connector-python==8.2.0
bcrypt==4.1.2
